fix dream island check and generally how Luna operates
thanks for daibunnie for figuring it out
https://github.com/Ixaruz/Luna-Archive-of-Dreams/issues/4

it now checks for the "SaveNPC Code" at the beginning of the SaveLand instead of the Town Name.

Luna now grabs the Town's Name from the first Player; sort of how NHSE does it.

I also moved all offsets/version specific sizes to util.h and defined them in util.cpp. should allow me to adapt to new ACNH versions in the future more easily.

Also making this a new release bc i felt like it.